+++ app/pagamento/page.tsx
@@ -1,199 +1,151 @@
-"use client"
+"use client";
 
-import {
-  Bell,
-  ShoppingCart,
-  ShieldCheck,
-  CreditCard,
-  QrCode,
-  Loader2,
-} from "lucide-react"
-import { useState } from "react"
-import { useRouter } from "next/navigation"
-import { MainHeader } from '../cliente/components/MainHeader';
+import { ShieldCheck, CreditCard, QrCode, Loader2 } from "lucide-react";
+import { useState } from "react";
+import { useRouter } from "next/navigation";
+import { MainHeader } from "../cliente/components/MainHeader";
 
 export default function PagamentoPage() {
-  const router = useRouter()
+  const router = useRouter();
 
-  const [metodo, setMetodo] = useState<"pix" | "cartao">("pix")
-  const [loading, setLoading] = useState(false)
-
-  const [form, setForm] = useState({
-    nome: "",
-    email: "",
-    documento: "",
-  })
-
-  const [errors, setErrors] = useState<any>({})
+  const [metodo, setMetodo] = useState<"pix" | "cartao">("pix");
+  const [loading, setLoading] = useState(false);
+  const [form, setForm] = useState({ nome: "", email: "", documento: "" });
+  const [errors, setErrors] = useState<any>({});
 
   function validar() {
-    const newErrors: any = {}
-
-    if (!form.nome) newErrors.nome = "Nome obrigatório"
-    if (!form.email.includes("@")) newErrors.email = "Email inválido"
-    if (form.documento.length < 11)
-      newErrors.documento = "CPF/CNPJ inválido"
-
-    setErrors(newErrors)
-    return Object.keys(newErrors).length === 0
+    const newErrors: any = {};
+    if (!form.nome) newErrors.nome = "Nome obrigatório";
+    if (!form.email.includes("@")) newErrors.email = "Email inválido";
+    if (form.documento.length < 11) newErrors.documento = "CPF/CNPJ inválido";
+    setErrors(newErrors);
+    return Object.keys(newErrors).length === 0;
   }
 
   function handleSubmit() {
-    if (!validar()) return
-
-    setLoading(true)
-
-    
+    if (!validar()) return;
+    setLoading(true);
     setTimeout(() => {
-      setLoading(false)
-
-    
-      router.push("/pagamento/status?status=sucesso")
-    }, 2000)
+      setLoading(false);
+      router.push("/pagamento/status?status=sucesso");
+    }, 2000);
   }
 
   return (
     <div className="min-h-screen bg-[#f6f8fb]">
+      <MainHeader />
 
-  <MainHeader />
+      <main className="mx-auto max-w-5xl px-4 py-10 md:px-6">
+        <h1 className="mb-8 text-3xl font-black text-gray-950">Pagamento</h1>
 
-      <div className="mx-auto max-w-6xl px-4 py-10 grid grid-cols-1 gap-6 lg:grid-cols-2">
+        <div className="grid grid-cols-1 gap-8 lg:grid-cols-2">
+          {/* Resumo do Pedido */}
+          <div className="rounded-3xl border border-gray-100 bg-white p-6 shadow-sm">
+            <h2 className="mb-5 text-sm font-black text-gray-950">Resumo do Pedido</h2>
 
-                <div className="rounded-3xl border border-gray-200 bg-white p-6">
-          <h2 className="text-lg font-semibold text-gray-800 mb-5">
-            Resumo do Pedido
-          </h2>
+            <div className="flex items-center gap-4 mb-6">
+              <div className="h-14 w-14 rounded-2xl bg-gray-50" />
+              <div>
+                <p className="font-black text-gray-950">Higienização de Sofá</p>
+                <p className="text-xs font-medium text-gray-400">ID: #849201</p>
+              </div>
+            </div>
 
-          <div className="flex items-center gap-4 mb-6">
-            <div className="h-14 w-14 rounded-xl bg-gray-100" />
-            <div>
-              <p className="font-medium text-gray-700">
-                Higienização de Sofá
-              </p>
-              <p className="text-sm text-gray-400">
-                ID: #849201
-              </p>
+            <div className="space-y-3 text-sm">
+              <div className="flex justify-between font-medium text-gray-500">
+                <span>Profissional</span>
+                <span className="font-bold text-gray-950">Carlos Silva</span>
+              </div>
+              <div className="flex justify-between font-medium text-gray-500">
+                <span>Data</span>
+                <span className="font-bold text-gray-950">11 Jun, 14:00</span>
+              </div>
+              <div className="flex justify-between font-medium text-gray-500">
+                <span>Taxa</span>
+                <span className="font-bold text-gray-950">R$ 0,00</span>
+              </div>
+            </div>
+
+            <div className="mt-6 border-t border-gray-100 pt-6 flex justify-between items-center">
+              <span className="font-black text-gray-950">Total</span>
+              <span className="text-2xl font-black text-yellow-600">R$ 250,00</span>
+            </div>
+
+            <div className="mt-6 flex gap-3 rounded-2xl bg-blue-50 p-4 text-xs font-bold text-blue-700">
+              <ShieldCheck size={18} className="shrink-0" />
+              Seu pagamento está protegido. O valor só será liberado após o serviço.
             </div>
           </div>
 
-          <div className="space-y-3 text-sm text-gray-600">
-            <div className="flex justify-between">
-              <span>Profissional</span>
-              <span className="text-gray-800">Carlos Silva</span>
+          {/* Formulário de Pagamento */}
+          <div className="rounded-3xl border border-gray-100 bg-white p-6 shadow-sm">
+            <h2 className="mb-5 text-sm font-black text-gray-950">Escolha o Método</h2>
+
+            <div className="flex gap-3 mb-6">
+              <button
+                onClick={() => setMetodo("pix")}
+                className={`flex-1 py-3 rounded-2xl border-2 flex items-center justify-center gap-2 text-sm font-black transition-all ${
+                  metodo === "pix" ? "border-yellow-400 bg-yellow-50 text-yellow-700" : "border-gray-100 text-gray-400 hover:border-gray-200"
+                }`}
+              >
+                <QrCode size={16} /> Pix
+              </button>
+              <button
+                onClick={() => setMetodo("cartao")}
+                className={`flex-1 py-3 rounded-2xl border-2 flex items-center justify-center gap-2 text-sm font-black transition-all ${
+                  metodo === "cartao" ? "border-yellow-400 bg-yellow-50 text-yellow-700" : "border-gray-100 text-gray-400 hover:border-gray-200"
+                }`}
+              >
+                <CreditCard size={16} /> Cartão
+              </button>
             </div>
 
-            <div className="flex justify-between">
-              <span>Data</span>
-              <span className="text-gray-800">11 Jun, 14:00</span>
+            <div className="space-y-4 mb-6">
+              <input
+                placeholder="Nome completo / Razão social"
+                className="w-full rounded-2xl border border-gray-200 px-5 py-3 text-sm font-medium focus:border-yellow-400 outline-none"
+                value={form.nome}
+                onChange={(e) => setForm({ ...form, nome: e.target.value })}
+              />
+              {errors.nome && <p className="text-[10px] font-bold text-red-500">{errors.nome}</p>}
+
+              <input
+                placeholder="Email"
+                className="w-full rounded-2xl border border-gray-200 px-5 py-3 text-sm font-medium focus:border-yellow-400 outline-none"
+                value={form.email}
+                onChange={(e) => setForm({ ...form, email: e.target.value })}
+              />
+              {errors.email && <p className="text-[10px] font-bold text-red-500">{errors.email}</p>}
+
+              <input
+                placeholder="CPF ou CNPJ"
+                className="w-full rounded-2xl border border-gray-200 px-5 py-3 text-sm font-medium focus:border-yellow-400 outline-none"
+                value={form.documento}
+                onChange={(e) => setForm({ ...form, documento: e.target.value })}
+              />
+              {errors.documento && <p className="text-[10px] font-bold text-red-500">{errors.documento}</p>}
             </div>
 
-            <div className="flex justify-between">
-              <span>Taxa</span>
-              <span className="text-gray-800">R$ 0,00</span>
-            </div>
-          </div>
-
-          <div className="mt-6 border-t pt-4 flex justify-between items-center">
-            <span className="font-medium text-gray-700">Total</span>
-            <span className="text-xl font-semibold text-yellow-600">
-              R$ 250,00
-            </span>
-          </div>
-
-          <div className="mt-6 flex gap-3 rounded-xl bg-blue-50 p-4 text-sm text-blue-700">
-            <ShieldCheck size={18} />
-            Seu pagamento está protegido. O valor só será liberado após o serviço.
+            <button
+              onClick={handleSubmit}
+              disabled={loading}
+              className="w-full rounded-2xl bg-yellow-400 py-4 font-black text-gray-950 transition-all hover:bg-yellow-500 active:scale-95 flex items-center justify-center gap-2"
+            >
+              {loading ? (
+                <>
+                  <Loader2 className="animate-spin" size={16} /> Processando...
+                </>
+              ) : (
+                "Confirmar e Pagar"
+              )}
+            </button>
           </div>
         </div>
+      </main>
 
-        <div className="rounded-3xl border border-gray-200 bg-white p-6">
-
-          <h2 className="text-lg font-semibold text-gray-800 mb-4">
-            Pagamento
-          </h2>
-
-          <div className="flex gap-3 mb-6">
-            <button
-              onClick={() => setMetodo("pix")}
-              className={`flex-1 py-2 rounded-xl border flex items-center justify-center gap-2 text-sm
-                ${metodo === "pix" ? "bg-yellow-50 border-yellow-400" : "border-gray-200"}
-              `}
-            >
-              <QrCode size={16} />
-              Pix
-            </button>
-
-            <button
-              onClick={() => setMetodo("cartao")}
-              className={`flex-1 py-2 rounded-xl border flex items-center justify-center gap-2 text-sm
-                ${metodo === "cartao" ? "bg-yellow-50 border-yellow-400" : "border-gray-200"}
-              `}
-            >
-              <CreditCard size={16} />
-              Cartão
-            </button>
-          </div>
-
-          <div className="space-y-4 mb-6">
-            <input
-              placeholder="Nome completo / Razão social"
-              className="w-full border rounded-xl px-4 py-2 text-sm"
-              value={form.nome}
-              onChange={(e) => setForm({ ...form, nome: e.target.value })}
-            />
-            {errors.nome && <p className="text-xs text-red-500">{errors.nome}</p>}
-
-            <input
-              placeholder="Email"
-              className="w-full border rounded-xl px-4 py-2 text-sm"
-              value={form.email}
-              onChange={(e) => setForm({ ...form, email: e.target.value })}
-            />
-            {errors.email && <p className="text-xs text-red-500">{errors.email}</p>}
-
-            <input
-              placeholder="CPF ou CNPJ"
-              className="w-full border rounded-xl px-4 py-2 text-sm"
-              value={form.documento}
-              onChange={(e) => setForm({ ...form, documento: e.target.value })}
-            />
-            {errors.documento && (
-              <p className="text-xs text-red-500">{errors.documento}</p>
-            )}
-          </div>
-
-          {metodo === "pix" && (
-            <div className="mb-6 text-sm text-gray-500">
-              Após confirmar, você receberá o QR Code para pagamento.
-            </div>
-          )}
-
-          {metodo === "cartao" && (
-            <div className="mb-6 text-sm text-gray-500">
-              Integração com cartão será feita no backend.
-            </div>
-          )}
-
-          <button
-            onClick={handleSubmit}
-            disabled={loading}
-            className="w-full rounded-full bg-yellow-400 py-3 font-medium hover:bg-yellow-500 flex items-center justify-center gap-2"
-          >
-            {loading ? (
-              <>
-                <Loader2 className="animate-spin" size={16} />
-                Processando...
-              </>
-            ) : (
-              "Confirmar e Pagar"
-            )}
-          </button>
-        </div>
-      </div>
-
-      <footer className="text-center text-xs text-gray-400 py-6">
+      <footer className="py-8 text-center text-xs font-medium text-gray-400">
         © 2026 Ajeitai - Todos os direitos reservados.
       </footer>
     </div>
-  )
+  );
 }
